Remove Babylon.js
