+++ env.d.ts
@@ -1,2 +1,3 @@
 /// <reference types="vite/client" />
 /// <reference types="unplugin-icons/types/vue" />
+/// <reference types="@webgpu/types" />
