+++ package.json
@@ -11,48 +11,41 @@
     "dev-bun": "bunx --bun vite --open",
     "build": "run-s build-only type-check",
     "preview": "vite preview",
-    "test:unit": "vitest",
     "build-only": "vite build",
     "type-check": "vue-tsc --build --force"
   },
   "dependencies": {
-    "@babylonjs/core": "^7.9.0",
-    "@babylonjs/loaders": "^7.9.0",
-    "@babylonjs/materials": "^7.9.0",
-    "@iconify-json/mdi": "^1.1.66",
+    "@iconify-json/mdi": "^1.1.67",
     "@stefnotch/typestef": "^1.0.3",
-    "@vueuse/core": "^10.10.0",
+    "@vueuse/core": "^10.11.0",
     "@zip.js/zip.js": "^2.7.45",
-    "monaco-editor": "^0.49.0",
+    "monaco-editor": "^0.50.0",
     "naive-ui": "^2.38.2",
     "pinia": "^2.1.7",
     "seemly": "^0.3.8",
-    "vue": "^3.4.27",
-    "vue-router": "^4.3.2",
+    "vue": "^3.4.31",
+    "vue-router": "^4.4.0",
     "webgpu-matrix": "^1.0.0",
-    "webgpu-utils": "^1.9.1",
+    "webgpu-utils": "^1.9.2",
     "zod": "^3.23.8",
-    "zod-to-json-schema": "^3.23.0"
+    "zod-to-json-schema": "^3.23.1"
   },
   "devDependencies": {
-    "@tsconfig/node20": "^20.1.2",
-    "@types/jsdom": "^21.1.7",
-    "@types/node": "^20.11.30",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/node": "^20.14.10",
     "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/test-utils": "^2.4.6",
     "@vue/tsconfig": "^0.5.1",
+    "@webgpu/types": "^0.1.44",
     "autoprefixer": "^10.4.19",
-    "jsdom": "^24.1.0",
-    "npm-run-all2": "^6.2.0",
-    "postcss": "^8.4.38",
-    "prettier": "3.3.0",
-    "tailwindcss": "^3.4.3",
-    "typescript": "~5.4.5",
+    "npm-run-all2": "^6.2.2",
+    "postcss": "^8.4.39",
+    "prettier": "3.3.2",
+    "tailwindcss": "^3.4.4",
+    "typescript": "~5.5.3",
     "unplugin-icons": "^0.19.0",
-    "unplugin-vue-components": "^0.27.0",
-    "vite": "^5.2.12",
-    "vitest": "^1.6.0",
-    "vue-tsc": "^2.0.19"
+    "unplugin-vue-components": "^0.27.2",
+    "vite": "^5.3.3",
+    "vue-tsc": "^2.0.26"
   },
   "browserslist": [
     "defaults"
