+++ src/scenes/scene-state.ts
