--- src/shaders/Common.wgsl
@@ -1,35 +1,0 @@
-struct Patch {
-  min: vec2<f32>,
-  max: vec2<f32>,
-};
-struct Patches {
-  patches_length: atomic<u32>,
-  patches_capacity: u32,
-  patches : array<Patch>,
-};
-struct PatchesRead { // Is currently needed, see https://github.com/gpuweb/gpuweb/discussions/4438
-  patches_length: u32, // Same size and alignment as atomic<u32>. Should be legal, right?
-  patches_capacity: u32,
-  patches : array<Patch>,
-};
-struct RenderBuffer {
-  patches_length: atomic<u32>,
-  patches_capacity: u32,
-  patches: array<Patch>,
-};
-struct RenderBufferRead {
-  patches_length: u32, // Same size as atomic<u32>
-  patches_capacity: u32,
-  patches: array<Patch>,
-};
-struct DispatchIndirectArgs { // From https://docs.rs/wgpu/latest/wgpu/util/struct.DispatchIndirectArgs.html
-  x: atomic<u32>,
-  y: u32,
-  z: u32,
-} 
-fn ceil_div(a: u32, b: u32) -> u32 { return (a + b - 1u) / b; }
-struct GlobalUBO {
-  iTime: f32,
-  iTimeDelta: f32,
-  iFrame: f32,
-};
