--- src/shaders/DefaultParametric.wgsl
@@ -1,4 +1,0 @@
-fn evaluateImage(input2: vec2f) -> vec3f {
-    let position = vec3(input2.x, input2.y, 0.0);
-    return position;
-}
