--- src/shaders/showcase/scene.json
@@ -1,160 +1,0 @@
-{
-  "$schema": "http://virtual/scene-schema.json",
-  "models": [
-    {
-      "type": "model",
-      "id": "74278bc2-e0d7-4c1e-9d42-1a7f35aca5d0",
-      "name": "Parametrisches Herz",
-      "position": [
-        5,
-        0,
-        0
-      ],
-      "rotation": [
-        0,
-        0,
-        0
-      ],
-      "scale": 5.300000000000001,
-      "parametricShader": "my-shader.vert.wgsl",
-      "material": {
-        "color": [
-          1,
-          0,
-          0
-        ],
-        "roughness": 0.5,
-        "metallic": 0.5,
-        "emissive": [
-          0,
-          0,
-          0
-        ]
-      }
-    },
-    {
-      "type": "model",
-      "id": "627684bf-d298-4841-b432-f4835e3ce617",
-      "name": "Parametrische Stachelkugel",
-      "position": [
-        -5,
-        0,
-        0
-      ],
-      "rotation": [
-        0,
-        0,
-        0
-      ],
-      "scale": 5,
-      "parametricShader": "spherehog-shader.vert.wgsl",
-      "material": {
-        "color": [
-          1,
-          0,
-          0
-        ],
-        "roughness": 0.5,
-        "metallic": 0.5,
-        "emissive": [
-          0,
-          0,
-          0
-        ]
-      }
-    },
-    {
-      "type": "model",
-      "id": "9cacec06-3dd8-48a9-a2cd-67fe3a28b220",
-      "name": "Parametrische Muschel",
-      "position": [
-        5,
-        20,
-        0
-      ],
-      "rotation": [
-        0,
-        0,
-        0
-      ],
-      "scale": 2,
-      "parametricShader": "seashell-shader.vert.wgsl",
-      "material": {
-        "color": [
-          1,
-          0,
-          0
-        ],
-        "roughness": 0.5,
-        "metallic": 0.5,
-        "emissive": [
-          0,
-          0,
-          0
-        ]
-      }
-    },
-    {
-      "type": "model",
-      "id": "56c0c0c9-fde6-4ac1-8514-bbe6bb7c6d34",
-      "name": "Parametrische Palmenblätter",
-      "position": [
-        -5,
-        20,
-        0
-      ],
-      "rotation": [
-        -1.5707963267948966,
-        0,
-        0
-      ],
-      "scale": 1,
-      "parametricShader": "palm-tree-shader.vert.wgsl",
-      "material": {
-        "color": [
-          1,
-          0,
-          0
-        ],
-        "roughness": 0.5,
-        "metallic": 0.5,
-        "emissive": [
-          0,
-          0,
-          0
-        ]
-      }
-    },
-    {
-      "type": "model",
-      "id": "35fa63a2-500f-4c7c-bc95-373833406542",
-      "name": "Parametrischer Palmenstamm",
-      "position": [
-        -5,
-        10,
-        0
-      ],
-      "rotation": [
-        0,
-        0,
-        0
-      ],
-      "scale": 1,
-      "parametricShader": "palm-tree-trunk-shader.vert.wgsl",
-      "material": {
-        "color": [
-          1,
-          0,
-          0
-        ],
-        "roughness": 0.5,
-        "metallic": 0.5,
-        "emissive": [
-          0,
-          0,
-          0
-        ]
-      }
-    }
-  ]
-}
